Move create instance step to common directory
